Extracts redux connection from PostList
This is a better approach because it decouples the view-level component logic, which is pure
React, from the mechanism for providing data to that component. By moving the Redux "connection"
into an entirely separate file, the component is agnostic to its data source: it could be rendered
by simply passing props like any other React component.

Another benefit of this change is later refactoring: if Redux later was swapped out for another
system, nobody needs to go refactor all the view components. Just delete the Redux connectors
and wire up the new system. Clean.